+++ src/mocks/browsers.ts
@@ -1,4 +1,4 @@
 import { setupWorker } from 'msw/browser';
-import { handlers } from './handlers';
+import { handlers } from './handlers/handlers';
 
 export const worker = setupWorker(...handlers);
